Add mean_cell_are and refactor grid_wrapper
