+++ model/driver/src/icon4py/model/driver/icon4py_driver.py
@@ -228,7 +228,7 @@ class TimeLoop:
             )
             self.solve_nonhydro.update_time_levels_for_velocity_tendencies(
                 solve_nonhydro_diagnostic_state,
-                at_first_substep=_is_first_substep(dyn_substep),
+                at_first_substep=self._is_first_substep(dyn_substep),
                 at_initial_timestep=self._is_first_step_in_simulation,
             )
             self.solve_nonhydro.time_step(
