+++ tools/src/icon4py/tools/py2fgen/wrappers/dycore_wrapper.py
@@ -6,18 +6,6 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
-# ICON4Py - ICON inspired code in Python and GT4Py
-#
-# Copyright (c) 2022, ETH Zurich and MeteoSwiss
-# All rights reserved.
-#
-# This file is free software: you can redistribute it and/or modify it under
-# the terms of the GNU General Public License as published by the
-# Free Software Foundation, either version 3 of the License, or any later
-# version. See the LICENSE.txt file at the top-level directory of this
-# distribution for a copy of the license or check <https://www.gnu.org/licenses/>.
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
 # type: ignore
 
 """
@@ -31,41 +19,38 @@ Fortran granule interfaces:
 
 import cProfile
 import pstats
+from typing import Optional
 
 import gt4py.next as gtx
 
 import icon4py.model.common.grid.states as grid_states
 from icon4py.model.atmosphere.dycore import dycore_states, solve_nonhydro
 from icon4py.model.common import dimension as dims, utils as common_utils
-from icon4py.model.common.grid import icon
-from icon4py.model.common.grid.icon import GlobalGridParams
+from icon4py.model.common.grid import icon as icon_grid
 from icon4py.model.common.grid.vertical import VerticalGrid, VerticalGridConfig
 from icon4py.model.common.states.prognostic_state import PrognosticState
 from icon4py.model.common.utils import data_allocation as data_alloc
 from icon4py.tools.common.logger import setup_logger
 from icon4py.tools.py2fgen.settings import backend, device
-from icon4py.tools.py2fgen.wrappers import common as wrapper_common
-from icon4py.tools.py2fgen.wrappers.wrapper_dimension import (
-    CellIndexDim,
-    EdgeIndexDim,
-    VertexIndexDim,
-)
+from icon4py.tools.py2fgen.wrappers import grid_wrapper
 
 
 logger = setup_logger(__name__)
 
-dycore_wrapper_state = {
-    "profiler": cProfile.Profile(),
-}
+# TODO(havogt): remove module global state
+profiler = cProfile.Profile()
+granule: Optional[solve_nonhydro.SolveNonhydro] = None
 
 
 def profile_enable():
-    dycore_wrapper_state["profiler"].enable()
+    global profiler
+    profiler.enable()
 
 
 def profile_disable():
-    dycore_wrapper_state["profiler"].disable()
-    stats = pstats.Stats(dycore_wrapper_state["profiler"])
+    global profiler
+    profiler.disable()
+    stats = pstats.Stats(profiler)
     stats.dump_stats(f"{__name__}.profile")
 
 
@@ -171,10 +156,11 @@ def solve_nh_init(
     lowest_layer_thickness: gtx.float64,
     model_top_height: gtx.float64,
     stretch_factor: gtx.float64,
+    mean_cell_area: gtx.float64,
     nflat_gradp: gtx.int32,
     num_levels: gtx.int32,
 ):
-    if not isinstance(dycore_wrapper_state["grid"], icon.IconGrid):
+    if not isinstance(grid_wrapper.grid, icon_grid.IconGrid):
         raise Exception("Need to initialise grid using grid_init before running solve_nh_init.")
 
     config = solve_nonhydro.NonHydrostaticConfig(
@@ -236,11 +222,11 @@ def solve_nh_init(
     )
 
     # datatest config CellParams
-    cell_geometry = grid_states.CellParams.from_global_num_cells(
+    cell_geometry = grid_states.CellParams(
         cell_center_lat=cell_center_lat,
         cell_center_lon=cell_center_lon,
         area=cell_areas,
-        global_num_cells=dycore_wrapper_state["grid"].global_properties.num_cells,
+        mean_cell_area=mean_cell_area,
         length_rescale_factor=1.0,
     )
 
@@ -322,8 +308,9 @@ def solve_nh_init(
         _min_index_flat_horizontal_grad_pressure=nflat_gradp,
     )
 
-    dycore_wrapper_state["granule"] = solve_nonhydro.SolveNonhydro(
-        grid=dycore_wrapper_state["grid"],
+    global granule
+    granule = solve_nonhydro.SolveNonhydro(
+        grid=grid_wrapper.grid,
         config=config,
         params=nonhydro_params,
         metric_state_nonhydro=metric_state_nonhydro,
@@ -333,6 +320,7 @@ def solve_nh_init(
         cell_geometry=cell_geometry,
         owner_mask=c_owner_mask,
         backend=backend,
+        exchange=grid_wrapper.exchange_runtime,
     )
 
 
@@ -426,7 +414,8 @@ def solve_nh_run(
     # adjust for Fortran indexes
     idyn_timestep = idyn_timestep - 1
 
-    dycore_wrapper_state["granule"].time_step(
+    global granule
+    granule.time_step(
         diagnostic_state_nh=diagnostic_state_nh,
         prognostic_states=prognostic_states,
         prep_adv=prep_adv,
@@ -436,61 +425,4 @@ def solve_nh_run(
         lprep_adv=lprep_adv,
         at_first_substep=idyn_timestep == 0,
         at_last_substep=idyn_timestep == (ndyn_substeps - 1),
-    )
-
-
-def grid_init(
-    cell_starts: gtx.Field[gtx.Dims[CellIndexDim], gtx.int32],
-    cell_ends: gtx.Field[gtx.Dims[CellIndexDim], gtx.int32],
-    vertex_starts: gtx.Field[gtx.Dims[VertexIndexDim], gtx.int32],
-    vertex_ends: gtx.Field[gtx.Dims[VertexIndexDim], gtx.int32],
-    edge_starts: gtx.Field[gtx.Dims[EdgeIndexDim], gtx.int32],
-    edge_ends: gtx.Field[gtx.Dims[EdgeIndexDim], gtx.int32],
-    c2e: gtx.Field[gtx.Dims[dims.CellDim, dims.C2EDim], gtx.int32],
-    e2c: gtx.Field[gtx.Dims[dims.EdgeDim, dims.E2CDim], gtx.int32],
-    c2e2c: gtx.Field[gtx.Dims[dims.CellDim, dims.C2E2CDim], gtx.int32],
-    e2c2e: gtx.Field[gtx.Dims[dims.EdgeDim, dims.E2C2EDim], gtx.int32],
-    e2v: gtx.Field[gtx.Dims[dims.EdgeDim, dims.E2VDim], gtx.int32],
-    v2e: gtx.Field[gtx.Dims[dims.VertexDim, dims.V2EDim], gtx.int32],
-    v2c: gtx.Field[gtx.Dims[dims.VertexDim, dims.V2CDim], gtx.int32],
-    e2c2v: gtx.Field[gtx.Dims[dims.EdgeDim, dims.E2C2VDim], gtx.int32],
-    c2v: gtx.Field[gtx.Dims[dims.CellDim, dims.C2VDim], gtx.int32],
-    global_root: gtx.int32,
-    global_level: gtx.int32,
-    num_vertices: gtx.int32,
-    num_cells: gtx.int32,
-    num_edges: gtx.int32,
-    vertical_size: gtx.int32,
-    limited_area: bool,
-):
-    # todo: write this logic into template.py
-    if isinstance(limited_area, int):
-        limited_area = bool(limited_area)
-
-    global_grid_params = GlobalGridParams(level=global_level, root=global_root)
-
-    dycore_wrapper_state["grid"] = wrapper_common.construct_icon_grid(
-        grid_id="icon_grid",
-        global_grid_params=global_grid_params,
-        num_vertices=num_vertices,
-        num_cells=num_cells,
-        num_edges=num_edges,
-        vertical_size=vertical_size,
-        limited_area=limited_area,
-        on_gpu=True if device == "GPU" else False,
-        cell_starts=cell_starts.ndarray,
-        cell_ends=cell_ends.ndarray,
-        vertex_starts=vertex_starts.ndarray,
-        vertex_ends=vertex_ends.ndarray,
-        edge_starts=edge_starts.ndarray,
-        edge_ends=edge_ends.ndarray,
-        c2e=c2e.ndarray,
-        e2c=e2c.ndarray,
-        c2e2c=c2e2c.ndarray,
-        e2c2e=e2c2e.ndarray,
-        e2v=e2v.ndarray,
-        v2e=v2e.ndarray,
-        v2c=v2c.ndarray,
-        e2c2v=e2c2v.ndarray,
-        c2v=c2v.ndarray,
     )
