+++ tools/tests/py2fgen/references/dycore/dycore.h
@@ -28,7 +28,8 @@ extern int solve_nh_run_wrapper(
     int grf_tend_vn_size_1, double *vn_ie, int vn_ie_size_0, int vn_ie_size_1,
     double *vt, int vt_size_0, int vt_size_1, double *mass_flx_me,
     int mass_flx_me_size_0, int mass_flx_me_size_1, double *mass_flx_ic,
-    int mass_flx_ic_size_0, int mass_flx_ic_size_1, double *vn_traj,
+    int mass_flx_ic_size_0, int mass_flx_ic_size_1, double *vol_flx_ic,
+    int vol_flx_ic_size_0, int vol_flx_ic_size_1, double *vn_traj,
     int vn_traj_size_0, int vn_traj_size_1, double dtime, int lprep_adv,
     int at_initial_timestep, double divdamp_fac_o2, double ndyn_substeps,
     int idyn_timestep);
@@ -117,17 +118,5 @@ extern int solve_nh_init_wrapper(
     double divdamp_fac2, double divdamp_fac3, double divdamp_fac4,
     double divdamp_z, double divdamp_z2, double divdamp_z3, double divdamp_z4,
     double lowest_layer_thickness, double model_top_height,
-    double stretch_factor, int nflat_gradp, int num_levels);
-extern int grid_init_wrapper(
-    int *cell_starts, int cell_starts_size_0, int *cell_ends,
-    int cell_ends_size_0, int *vertex_starts, int vertex_starts_size_0,
-    int *vertex_ends, int vertex_ends_size_0, int *edge_starts,
-    int edge_starts_size_0, int *edge_ends, int edge_ends_size_0, int *c2e,
-    int c2e_size_0, int c2e_size_1, int *e2c, int e2c_size_0, int e2c_size_1,
-    int *c2e2c, int c2e2c_size_0, int c2e2c_size_1, int *e2c2e,
-    int e2c2e_size_0, int e2c2e_size_1, int *e2v, int e2v_size_0,
-    int e2v_size_1, int *v2e, int v2e_size_0, int v2e_size_1, int *v2c,
-    int v2c_size_0, int v2c_size_1, int *e2c2v, int e2c2v_size_0,
-    int e2c2v_size_1, int *c2v, int c2v_size_0, int c2v_size_1, int global_root,
-    int global_level, int num_vertices, int num_cells, int num_edges,
-    int vertical_size, int limited_area);
+    double stretch_factor, double mean_cell_area, int nflat_gradp,
+    int num_levels);
