+++ tools/tests/py2fgen/references/dycore/dycore.py
@@ -18,27 +18,20 @@ logging.info(cp.show_config())
 # embedded function imports
 from icon4py.tools.py2fgen.wrappers.dycore_wrapper import solve_nh_run
 from icon4py.tools.py2fgen.wrappers.dycore_wrapper import solve_nh_init
-from icon4py.tools.py2fgen.wrappers.dycore_wrapper import grid_init
 
 
 C2E = gtx.Dimension("C2E", kind=gtx.DimensionKind.LOCAL)
-C2E2C = gtx.Dimension("C2E2C", kind=gtx.DimensionKind.LOCAL)
 C2E2CO = gtx.Dimension("C2E2CO", kind=gtx.DimensionKind.LOCAL)
-C2V = gtx.Dimension("C2V", kind=gtx.DimensionKind.LOCAL)
 Cell = gtx.Dimension("Cell", kind=gtx.DimensionKind.HORIZONTAL)
-CellIndex = gtx.Dimension("CellIndex", kind=gtx.DimensionKind.HORIZONTAL)
 E2C = gtx.Dimension("E2C", kind=gtx.DimensionKind.LOCAL)
 E2C2E = gtx.Dimension("E2C2E", kind=gtx.DimensionKind.LOCAL)
 E2C2EO = gtx.Dimension("E2C2EO", kind=gtx.DimensionKind.LOCAL)
 E2C2V = gtx.Dimension("E2C2V", kind=gtx.DimensionKind.LOCAL)
-E2V = gtx.Dimension("E2V", kind=gtx.DimensionKind.LOCAL)
 Edge = gtx.Dimension("Edge", kind=gtx.DimensionKind.HORIZONTAL)
-EdgeIndex = gtx.Dimension("EdgeIndex", kind=gtx.DimensionKind.HORIZONTAL)
 K = gtx.Dimension("K", kind=gtx.DimensionKind.VERTICAL)
 V2C = gtx.Dimension("V2C", kind=gtx.DimensionKind.LOCAL)
 V2E = gtx.Dimension("V2E", kind=gtx.DimensionKind.LOCAL)
 Vertex = gtx.Dimension("Vertex", kind=gtx.DimensionKind.HORIZONTAL)
-VertexIndex = gtx.Dimension("VertexIndex", kind=gtx.DimensionKind.HORIZONTAL)
 
 
 @ffi.def_extern()
@@ -133,6 +126,9 @@ def solve_nh_run_wrapper(
     mass_flx_ic,
     mass_flx_ic_size_0,
     mass_flx_ic_size_1,
+    vol_flx_ic,
+    vol_flx_ic_size_0,
+    vol_flx_ic_size_1,
     vn_traj,
     vn_traj_size_0,
     vn_traj_size_1,
@@ -383,6 +379,15 @@ def solve_nh_run_wrapper(
             False,
         )
 
+        vol_flx_ic = wrapper_utils.as_field(
+            ffi,
+            xp,
+            vol_flx_ic,
+            ts.ScalarKind.FLOAT64,
+            {Cell: vol_flx_ic_size_0, K: vol_flx_ic_size_1},
+            False,
+        )
+
         vn_traj = wrapper_utils.as_field(
             ffi,
             xp,
@@ -429,6 +434,7 @@ def solve_nh_run_wrapper(
             vt,
             mass_flx_me,
             mass_flx_ic,
+            vol_flx_ic,
             vn_traj,
             dtime,
             lprep_adv,
@@ -694,6 +700,15 @@ def solve_nh_run_wrapper(
         )
         logging.debug(msg)
 
+        msg = "shape of vol_flx_ic after computation = %s" % str(
+            vol_flx_ic.shape if vol_flx_ic is not None else "None"
+        )
+        logging.debug(msg)
+        msg = "vol_flx_ic after computation: %s" % str(
+            vol_flx_ic.ndarray if vol_flx_ic is not None else "None"
+        )
+        logging.debug(msg)
+
         msg = "shape of vn_traj after computation = %s" % str(
             vn_traj.shape if vn_traj is not None else "None"
         )
@@ -939,6 +954,7 @@ def solve_nh_init_wrapper(
     lowest_layer_thickness,
     model_top_height,
     stretch_factor,
+    mean_cell_area,
     nflat_gradp,
     num_levels,
 ):
@@ -1614,6 +1630,7 @@ def solve_nh_init_wrapper(
             lowest_layer_thickness,
             model_top_height,
             stretch_factor,
+            mean_cell_area,
             nflat_gradp,
             num_levels,
         )
@@ -2271,264 +2288,6 @@ def solve_nh_init_wrapper(
         msg = "primal_normal_y after computation: %s" % str(
             primal_normal_y.ndarray if primal_normal_y is not None else "None"
         )
-        logging.debug(msg)
-
-        logging.critical("Python Execution Context End")
-
-    except Exception as e:
-        logging.exception(f"A Python error occurred: {e}")
-        return 1
-
-    return 0
-
-
-@ffi.def_extern()
-def grid_init_wrapper(
-    cell_starts,
-    cell_starts_size_0,
-    cell_ends,
-    cell_ends_size_0,
-    vertex_starts,
-    vertex_starts_size_0,
-    vertex_ends,
-    vertex_ends_size_0,
-    edge_starts,
-    edge_starts_size_0,
-    edge_ends,
-    edge_ends_size_0,
-    c2e,
-    c2e_size_0,
-    c2e_size_1,
-    e2c,
-    e2c_size_0,
-    e2c_size_1,
-    c2e2c,
-    c2e2c_size_0,
-    c2e2c_size_1,
-    e2c2e,
-    e2c2e_size_0,
-    e2c2e_size_1,
-    e2v,
-    e2v_size_0,
-    e2v_size_1,
-    v2e,
-    v2e_size_0,
-    v2e_size_1,
-    v2c,
-    v2c_size_0,
-    v2c_size_1,
-    e2c2v,
-    e2c2v_size_0,
-    e2c2v_size_1,
-    c2v,
-    c2v_size_0,
-    c2v_size_1,
-    global_root,
-    global_level,
-    num_vertices,
-    num_cells,
-    num_edges,
-    vertical_size,
-    limited_area,
-):
-    try:
-        logging.info("Python Execution Context Start")
-
-        # Convert ptr to GT4Py fields
-
-        cell_starts = wrapper_utils.as_field(
-            ffi, xp, cell_starts, ts.ScalarKind.INT32, {CellIndex: cell_starts_size_0}, False
-        )
-
-        cell_ends = wrapper_utils.as_field(
-            ffi, xp, cell_ends, ts.ScalarKind.INT32, {CellIndex: cell_ends_size_0}, False
-        )
-
-        vertex_starts = wrapper_utils.as_field(
-            ffi, xp, vertex_starts, ts.ScalarKind.INT32, {VertexIndex: vertex_starts_size_0}, False
-        )
-
-        vertex_ends = wrapper_utils.as_field(
-            ffi, xp, vertex_ends, ts.ScalarKind.INT32, {VertexIndex: vertex_ends_size_0}, False
-        )
-
-        edge_starts = wrapper_utils.as_field(
-            ffi, xp, edge_starts, ts.ScalarKind.INT32, {EdgeIndex: edge_starts_size_0}, False
-        )
-
-        edge_ends = wrapper_utils.as_field(
-            ffi, xp, edge_ends, ts.ScalarKind.INT32, {EdgeIndex: edge_ends_size_0}, False
-        )
-
-        c2e = wrapper_utils.as_field(
-            ffi, xp, c2e, ts.ScalarKind.INT32, {Cell: c2e_size_0, C2E: c2e_size_1}, False
-        )
-
-        e2c = wrapper_utils.as_field(
-            ffi, xp, e2c, ts.ScalarKind.INT32, {Edge: e2c_size_0, E2C: e2c_size_1}, False
-        )
-
-        c2e2c = wrapper_utils.as_field(
-            ffi, xp, c2e2c, ts.ScalarKind.INT32, {Cell: c2e2c_size_0, C2E2C: c2e2c_size_1}, False
-        )
-
-        e2c2e = wrapper_utils.as_field(
-            ffi, xp, e2c2e, ts.ScalarKind.INT32, {Edge: e2c2e_size_0, E2C2E: e2c2e_size_1}, False
-        )
-
-        e2v = wrapper_utils.as_field(
-            ffi, xp, e2v, ts.ScalarKind.INT32, {Edge: e2v_size_0, E2V: e2v_size_1}, False
-        )
-
-        v2e = wrapper_utils.as_field(
-            ffi, xp, v2e, ts.ScalarKind.INT32, {Vertex: v2e_size_0, V2E: v2e_size_1}, False
-        )
-
-        v2c = wrapper_utils.as_field(
-            ffi, xp, v2c, ts.ScalarKind.INT32, {Vertex: v2c_size_0, V2C: v2c_size_1}, False
-        )
-
-        e2c2v = wrapper_utils.as_field(
-            ffi, xp, e2c2v, ts.ScalarKind.INT32, {Edge: e2c2v_size_0, E2C2V: e2c2v_size_1}, False
-        )
-
-        c2v = wrapper_utils.as_field(
-            ffi, xp, c2v, ts.ScalarKind.INT32, {Cell: c2v_size_0, C2V: c2v_size_1}, False
-        )
-
-        assert isinstance(limited_area, int)
-        limited_area = limited_area != 0
-
-        grid_init(
-            cell_starts,
-            cell_ends,
-            vertex_starts,
-            vertex_ends,
-            edge_starts,
-            edge_ends,
-            c2e,
-            e2c,
-            c2e2c,
-            e2c2e,
-            e2v,
-            v2e,
-            v2c,
-            e2c2v,
-            c2v,
-            global_root,
-            global_level,
-            num_vertices,
-            num_cells,
-            num_edges,
-            vertical_size,
-            limited_area,
-        )
-
-        # debug info
-
-        msg = "shape of cell_starts after computation = %s" % str(
-            cell_starts.shape if cell_starts is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "cell_starts after computation: %s" % str(
-            cell_starts.ndarray if cell_starts is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of cell_ends after computation = %s" % str(
-            cell_ends.shape if cell_ends is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "cell_ends after computation: %s" % str(
-            cell_ends.ndarray if cell_ends is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of vertex_starts after computation = %s" % str(
-            vertex_starts.shape if vertex_starts is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "vertex_starts after computation: %s" % str(
-            vertex_starts.ndarray if vertex_starts is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of vertex_ends after computation = %s" % str(
-            vertex_ends.shape if vertex_ends is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "vertex_ends after computation: %s" % str(
-            vertex_ends.ndarray if vertex_ends is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of edge_starts after computation = %s" % str(
-            edge_starts.shape if edge_starts is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "edge_starts after computation: %s" % str(
-            edge_starts.ndarray if edge_starts is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of edge_ends after computation = %s" % str(
-            edge_ends.shape if edge_ends is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "edge_ends after computation: %s" % str(
-            edge_ends.ndarray if edge_ends is not None else "None"
-        )
-        logging.debug(msg)
-
-        msg = "shape of c2e after computation = %s" % str(c2e.shape if c2e is not None else "None")
-        logging.debug(msg)
-        msg = "c2e after computation: %s" % str(c2e.ndarray if c2e is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of e2c after computation = %s" % str(e2c.shape if e2c is not None else "None")
-        logging.debug(msg)
-        msg = "e2c after computation: %s" % str(e2c.ndarray if e2c is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of c2e2c after computation = %s" % str(
-            c2e2c.shape if c2e2c is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "c2e2c after computation: %s" % str(c2e2c.ndarray if c2e2c is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of e2c2e after computation = %s" % str(
-            e2c2e.shape if e2c2e is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "e2c2e after computation: %s" % str(e2c2e.ndarray if e2c2e is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of e2v after computation = %s" % str(e2v.shape if e2v is not None else "None")
-        logging.debug(msg)
-        msg = "e2v after computation: %s" % str(e2v.ndarray if e2v is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of v2e after computation = %s" % str(v2e.shape if v2e is not None else "None")
-        logging.debug(msg)
-        msg = "v2e after computation: %s" % str(v2e.ndarray if v2e is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of v2c after computation = %s" % str(v2c.shape if v2c is not None else "None")
-        logging.debug(msg)
-        msg = "v2c after computation: %s" % str(v2c.ndarray if v2c is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of e2c2v after computation = %s" % str(
-            e2c2v.shape if e2c2v is not None else "None"
-        )
-        logging.debug(msg)
-        msg = "e2c2v after computation: %s" % str(e2c2v.ndarray if e2c2v is not None else "None")
-        logging.debug(msg)
-
-        msg = "shape of c2v after computation = %s" % str(c2v.shape if c2v is not None else "None")
-        logging.debug(msg)
-        msg = "c2v after computation: %s" % str(c2v.ndarray if c2v is not None else "None")
         logging.debug(msg)
 
         logging.critical("Python Execution Context End")
